Extract core
