+++ YewCone.AudiobookGenerator.Core/YewCone.AudiobookGenerator.Core.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>net9.0-windows</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" />
+    <PackageReference Include="Microsoft.Playwright"/>
+    <PackageReference Include="System.Speech"/>
+    <PackageReference Include="VersOne.Epub"/>
+    <PackageReference Include="FFMpegCore"/>
+    <PackageReference Include="TagLibSharp"/>
+  </ItemGroup>
+</Project>
