+++ YewCone.AudiobookGenerator.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.6.33706.43
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YewCone.AudiobookGenerator", "AudiobookGenerator\YewCone.AudiobookGenerator.csproj", "{ED68EF4A-CB8A-4019-A355-F494AFCC5EDA}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{F42EB3EE-76A4-4B37-B840-E5FFB5FF162B}"
 	ProjectSection(SolutionItems) = preProject
 		.gitattributes = .gitattributes
@@ -16,16 +14,24 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "YewCone.AudiobookGenerator.Core", "YewCone.AudiobookGenerator.Core\YewCone.AudiobookGenerator.Core.csproj", "{8E81DB05-1456-4C9C-B8C5-84DB0488A80C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "YewCone.AudiobookGenerator.Console", "YewCone.AudiobookGenerator.Console\YewCone.AudiobookGenerator.Console.csproj", "{876376BA-9FC6-4F65-8EFB-4E8E75B183C3}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{ED68EF4A-CB8A-4019-A355-F494AFCC5EDA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ED68EF4A-CB8A-4019-A355-F494AFCC5EDA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ED68EF4A-CB8A-4019-A355-F494AFCC5EDA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ED68EF4A-CB8A-4019-A355-F494AFCC5EDA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8E81DB05-1456-4C9C-B8C5-84DB0488A80C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8E81DB05-1456-4C9C-B8C5-84DB0488A80C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8E81DB05-1456-4C9C-B8C5-84DB0488A80C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8E81DB05-1456-4C9C-B8C5-84DB0488A80C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{876376BA-9FC6-4F65-8EFB-4E8E75B183C3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{876376BA-9FC6-4F65-8EFB-4E8E75B183C3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{876376BA-9FC6-4F65-8EFB-4E8E75B183C3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{876376BA-9FC6-4F65-8EFB-4E8E75B183C3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
